[TEST] charconv: split charconv test case
