+++ test/unit/std/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectories()
 
-seqan3_test(charconv_test.cpp)
+seqan3_test(charconv_int_test.cpp)
+seqan3_test(charconv_float_test.cpp)
 seqan3_test(ranges_test.cpp)
